Some of the comment for internam logic

+++ TodoList.Api-master/TodoList.Api/Controllers/TodoItemsController.cs
@@ -38,7 +38,11 @@ namespace TodoList.Api.Controllers
                 return StatusCode(500, "Internal server error");
             }
         }
-
+        /// <summary>
+        /// This is a get request where we are trying to Fetch Item from the list via existing GUID in a list
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns>It will return with the specific item from the list</returns>
         [HttpGet("{id}")]
         public async Task<IActionResult> GetTodoItem(Guid id)
         {
@@ -57,7 +61,15 @@ namespace TodoList.Api.Controllers
                 return StatusCode(500, "Internal server error");
             }
         }
-
+        /// <summary>
+        /// This will allow to ADD the Item in the list via HTTP post request
+        /// <Validation>
+        /// It will handle the bad request if the Item already exists.
+        /// It will handle if the item we are trying to post with empty list
+        /// </Validation>
+        /// </summary>
+        /// <param name="todoItemDTO"></param>
+        /// <returns></returns>
         [HttpPost]
         public async Task<IActionResult> CreateTodoItem([FromBody] TodoItemDTO todoItemDTO)
         {
@@ -85,7 +97,16 @@ namespace TodoList.Api.Controllers
                 return StatusCode(500, "Internal server error");
             }
         }
-
+        /// <summary>
+        /// This will allow to update the TODO Item list via GUID provided.
+        /// </summary>
+        /// <Validation>
+        /// It provide with user friendly message if someone tring to update the TODO item visa wrong ID whcich is not exists in a list
+        /// It will handle if the item we are trying to post with empty list
+        /// </Validation>
+        /// <param name="id"></param>
+        /// <param name="todoItemDTO"></param>
+        /// <returns></returns>
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateTodoItem(Guid id, [FromBody] TodoItemDTO todoItemDTO)
         {
@@ -111,12 +132,26 @@ namespace TodoList.Api.Controllers
                 return StatusCode(500, "Internal server error");
             }
         }
-
+       
+        /// <summary>
+        /// This will allow to delete the TODO Item list via GUID provided.
+        /// </summary>
+        /// <Validation>
+        /// It provide with user friendly message if someone tring to elete the TODO item visa wrong ID whcich is not exists in a list
+        /// </Validation>
+        /// <param name="id"></param>
+        /// <returns></returns>
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteTodoItem(Guid id)
         {
             try
             {
+                var existingItem = await _repository.GetByIdAsync(id);
+                if (existingItem == null)
+                {
+                    return BadRequest("ID not found");  // Returns ID not found use
+                }
+
                 await _repository.DeleteAsync(id);
                 return NoContent();
             }
